refactor: 清理 updater 模块重复代码与冗余 I/O
- 提取 sha256_hex 到 mod.rs，消除 state.rs 和 helper.rs 的重复实现
- 复用 check.rs 的 env_manifest_path，移除 download.rs 的副本
- 复用 platform.rs 的 find_macos_app_bundle，移除 install.rs 的副本
- 统一 debug_log! 宏定义到 mod.rs，各子模块委托调用
- 简化 matches_install_kind 为单行等值比较
- finalize_update_check 复用已加载的状态，避免重复读取 state.json
- 移除 state/settings/install 中冗余的 ensure_dirs() 调用
